[DRAFT 509735] Display activities in the Sirius aird editor
Add a new plug-in which provides additional pages to the Sirius
session editor displaying the same pages as in the Activity Explorer,
using the information from the static extension point
configuration (for now).

The current state is very minimal, but this integration is the final
goal, so this will serve to drive the changes needed in the AE code to
improve the integration piece by piece.

Bug: 509735
Change-Id: I98df194728357a004ff11a53b3c1db04ecb42847
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>